Add ServiceIdentifier, rename SharedIdentifier to PlatformIdentifier
This is needed to detect when there are services that are registered to a platform where IP or port can change without recreating the platform.
This can happen when for example, a docker container with a dynamic port-mapping is restarted or when a docker container with an internal network is restarted.
Without live-watching (disabled or dropped events), we are not able to detect that case and leave old services in the registry