Improve reregister containers test to catch race
There is a race between the marker service appearing and the other service being replaced, it depends on the
order that the services are diff'ed in. If the marker service appears before the service under test is replaced,
we may see a failing test because the new service is not yet registered or the old one is not yet deregistered.

By waiting for 2 anti-entropy cycles, we avoid this issue because handling the service under test happens in the
first anti-entropy cycle.