Wait for complete registration before checking multiple services
The serviceExists() check waits for a first service instance to appear,
not until all service instances have appeared.

Sometimes, this is not a problem, since everything will already be
registered in consul by the time the catalog is fetched. Other times,
not all service instances have yet been registered, resulting in the
test failing.