Use checked exceptions for failures in registry
All registry operations should be regarded as prone to failure.
For example, they may fail because the registry does not accept the configuration set
on a certain service. In those cases, it is best not to stop registering
other services.

Unchecked exceptions lead to the pattern of not checking for failure,
especially if failure does not happen often.
Therefore, checked exceptions will require failures to be handled at
some point, in this case it would be preferrably in a place that does
not prevent the registration loop from making progress.